% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikehood.r
\name{calculateBaseline}
\alias{calculateBaseline}
\title{Calculate Age-Specific and Cumulative Non-Carrier Penetrance}
\usage{
calculateBaseline(cancer_type, gene, race, type, db)
}
\arguments{
\item{SEER_baseline}{Numeric vector, the baseline risk for the general population by age.}

\item{alpha}{Numeric, the shape parameter of the Weibull distribution for the genetic risk.}

\item{beta}{Numeric, the scale parameter of the Weibull distribution for the genetic risk.}

\item{delta}{Numeric, the shift parameter for the age in the Weibull risk calculation.}

\item{gamma}{Numeric, the adjustment factor for the genetic risk.}

\item{af}{Numeric, the allele frequency of the risk allele in the population.}

\item{max_age}{Integer, the maximum age to calculate the penetrance for.}
}
\value{
A list containing two elements: 'yearlyProb' with the yearly probability of not developing the disease,
and 'cumulativeProb' with the cumulative probability up to each age.
}
\description{
This function calculates the age-specific and cumulative non-carrier penetrance based on SEER baseline
data, genetic risk parameters, and allele frequencies.
}
