% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikehood.r
\name{calculateNCPen}
\alias{calculateNCPen}
\title{Calculate Age-Specific Non-Carrier Penetrance}
\usage{
calculateNCPen(SEER_baseline, alpha, beta, delta, gamma, af, max_age)
}
\arguments{
\item{SEER_baseline}{Numeric, the baseline penetrance derived from SEER data for the general population without considering genetic risk factors.}

\item{af}{Numeric, the allele frequency of the risk allele in the population.}

\item{carrierPenetrances}{Numeric vector, the penetrances for carriers at different ages and possibly by sex, assuming that the structure aligns with the age and possibly sex.}
}
\value{
Numeric, the calculated age-specific non-carrier penetrance, adjusted for genetic risk.
}
\description{
This function calculates the age-specific non-carrier penetrance based on SEER baseline
data, penetrances for carriers, and allele frequencies. It is designed to adjust penetrance estimates
for genetic testing by incorporating the genetic risk attributable to specified alleles.
}
\examples{
SEER_baseline <- 0.05 # Example baseline penetrance
carrierPenetrances <- c(0.1, 0.2, 0.3) # Example carrier penetrances
af <- 0.01 # Example allele frequency
calculateNCPen(SEER_baseline, carrierPenetrances, af)

}
