% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenEstimMain.R
\name{mhChain}
\alias{mhChain}
\title{Execution of a Single Chain in Metropolis-Hastings for Cancer Risk Estimation}
\usage{
mhChain(
  seed,
  n_iter,
  burn_in,
  chain_id,
  ncores,
  data,
  max_age,
  baseline_data,
  prior_distributions,
  af,
  median_max,
  max_penetrance,
  BaselineNC,
  var,
  ageImputation,
  removeProband
)
}
\arguments{
\item{seed}{Seed value for random number generation.}

\item{n_iter}{Number of iterations for the chain.}

\item{burn_in}{Fraction of iterations to discard as burn-in (0 to 1).}

\item{chain_id}{Identifier for the chain.}

\item{data}{Transformed dataframe in the format compatible with clipp.}

\item{max_age}{Maximum age to be considered.}

\item{baseline_data}{Data containing baseline risk estimates in manual input format.}

\item{prior_distributions}{List of the parameters for the distributions of the proposal, including asymptote, threshold, median, and first quartile distributions.}

\item{af}{Allele frequency for the risk allele.}

\item{median_max}{Boolean indicating whether to use baseline median or max_age as an upper bound for the median proposal. Defaults to TRUE, i.e., using the baseline median.}

\item{max_penetrance}{Maximum penetrance considered for analysis.}

\item{BaselineNC}{Boolean indicating if non-carrier penetrance is assumed to be the Baseline penetrance.}

\item{var}{Variance-covariance matrix used in the proposal distribution.}

\item{ageImputation}{Boolean indicating whether to impute ages for missing data.}
}
\value{
A list containing samples, log likelihoods, acceptance ratio, and rejection rate for each iteration.
}
\description{
Performs a single chain execution in the Metropolis-Hastings algorithm for Bayesian inference,
specifically tailored for cancer risk estimation. It estimates parameters related to cancer penetrance
based on family data, genetic information, and baseline database estimates.
}
