% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikehood.r
\name{mhLogLikelihood_clipp}
\alias{mhLogLikelihood_clipp}
\title{Calculate Log Likelihood using clipp Package}
\usage{
mhLogLikelihood_clipp(
  paras,
  families,
  max_age,
  seer_data,
  af,
  BaselineNC,
  ncores
)
}
\arguments{
\item{paras}{A vector of parameters (proposals)}

\item{families}{The pedigree data for families.}

\item{max_age}{The maximum age considered.}

\item{af}{Allele frequency.}

\item{cancer_type}{The type of cancer.}

\item{db}{The data object.}
}
\value{
The log likelihood value.
}
\description{
This function calculates the log likelihood using the clipp package for a set of parameters and data.
}
\examples{
# Calculate log likelihood
log_likelihood <- mhLogLikelihood_clipp(parameters, pedigree_data, 80, "Lung Cancer", data_object, 0.2)
}
\references{
Details about the clipp package and methods can be found in the package documentation.
}
