% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenEstimMain.R
\name{PenEstim}
\alias{PenEstim}
\title{Bayesian Inference using Independent Metropolis-Hastings for Penetrance Estimation}
\usage{
PenEstim(
  data,
  n_chains = 1,
  n_iter_per_chain = 10000,
  ncores = 6,
  max_age = 94,
  baseline_data = PPP::PanelPRODatabase$AlleleFrequency[paste0("MLH1", "_anyPV"),
    "nonAJ"],
  removeProband = FALSE,
  ageImputation = FALSE,
  median_max = TRUE,
  BaselineNC = TRUE,
  var = c(0.1, 0.1, 2, 2, 5, 5, 5, 5),
  burn_in = 0,
  thinning_factor = 1,
  distribution_data = distribution_data_default,
  af = 1e-04,
  max_penetrance = 1,
  sample_size = NULL,
  ratio = NULL,
  prior_params = prior_params_default,
  risk_proportion = risk_proportion_default,
  summary_stats = TRUE,
  rejection_rates = TRUE,
  density_plots = TRUE,
  penetrance_plot = TRUE,
  probCI = 0.95
)
}
\arguments{
\item{data}{The pedigree data in the required format. A data frame of
family history information with the following columns.
\itemize{
\item \code{ID}: A numeric value; ID for each individual. There should not be any duplicated entries.
\item \code{Sex}: A numeric value; \code{0} for female and \code{1} for male. Missing entries are not currently supported.
\item \code{MotherID}: A numeric value; unique ID for someone's mother.
\item \code{FatherID}: A numeric value; unique ID for someone's father.
\item \code{isProband}: A numeric value; \code{1} if someone is a proband, \code{0} otherwise.
\item \code{CurAge}: A numeric value; the age of censoring (current age if the person is alive or age of death if the person is dead). Ages ranging from \code{1} to \code{94} are allowed.
\item \code{isAff}: A numeric value; the affection status of cancer, encoded as \code{1} if the individual was diagnosed, \code{0} otherwise. Missing entries are not currently supported.
\item \code{Age}: A numeric value; the age of diagnosis for cancer, encoded as \code{NA} if the individual was not diagnosed. Ages ranging from \code{1} to \code{94} are allowed.
\item \code{isDead}: A numeric value; \code{1} if someone is dead, \code{0} otherwise. Missing entries are assumed to be \code{0}.
\item \code{geno}:  Column for germline testing results or tumor marker testing results. Positive results should be coded as \code{1},
negative results should be coded as \code{0}, and unknown results should be coded as \code{NA} or left empty.
\item Optional: \code{race}: A character string; expected values are \code{"All_Races"}, \code{"AIAN"}, \code{"Asian"}, \code{"Black"}, \code{"White"}, \code{"Hispanic"}, \code{"WH"}, and \code{"WNH"}.
\item Optional: \code{Ancestry}: A character string; expected values are \code{"AJ"}, \code{"nonAJ"}, and \code{"Italian"}.
\item \code{Twins}: A numeric value; \code{0} for non-identical/single births, \code{1} for the first set of identical twins/multiple births in the family, \code{2} for the second set, etc.
}}

\item{n_chains}{Number of chains for parallel computation.}

\item{n_iter_per_chain}{Number of iterations for each chain.}

\item{max_age}{Maximum age considered for analysis, default is 94.}

\item{baseline_data}{Data for the baseline risk estimates (probability of developing cancer), for instance population-level risk from a cancer registry.}

\item{removeProband}{Logical, indicating whether to remove probands from the analysis (default is FALSE).}

\item{median_max}{Boolean indicating whether to use baseline median age or max_age as an upper bound for the median proposal. Defaults to TRUE.}

\item{BaselineNC}{Boolean indicating that the non-carrier penetrance is assumed to be the Baseline penetrance. Default is TRUE.}

\item{burn_in}{Fraction of results to discard as burn-in (0 to 1). Default is 0 (no burn-in).}

\item{thinning_factor}{Factor by which to thin the results, default is 1 (no thinning).}

\item{distribution_data}{Data for generating prior distributions.}

\item{af}{Allele frequency for risk allele, default is 0.0001.}

\item{max_penetrance}{Maximum penetrance considered for analysis, default is 1.}

\item{sample_size}{Optional sample size for distribution generation.}

\item{ratio}{Optional ratio parameter for distribution generation.}

\item{prior_params}{Parameters for prior distributions.}

\item{risk_proportion}{Proportion of risk for distribution generation.}

\item{summary_stats}{Boolean to include summary statistics in the output.}

\item{rejection_rates}{Boolean to include rejection rates in the output.}

\item{density_plots}{Boolean to include density plots in the output.}

\item{penetrance_plot}{Boolean to include penetrance plots in the output.}

\item{probCI}{Probability level for confidence intervals in penetrance plots (default is 0.95).}

\item{sex}{Option to run the estimation for each sex separately. The default is estimation of one penetrance curve for both sexes (sex = "NA").}
}
\value{
A list containing combined results from all chains, along with optional statistics and plots.
}
\description{
This function implements the Independent Metropolis-Hastings algorithm for Bayesian
penetrance estimation of cancer risk. It utilizes parallel computing to run multiple
chains and provides various options for analyzing and visualizing the results.
}
