% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{plot_penetrance}
\alias{plot_penetrance}
\title{Plot Weibull Distribution with Credible Intervals}
\usage{
plot_penetrance(data, prob, max_age, sex = "NA")
}
\arguments{
\item{data}{A list with combined results from MCMC.}

\item{prob}{The probability for the credible interval (between 0 and 1).}
}
\description{
Plot Weibull Distribution with Credible Intervals
}
\examples{
plot_weibull_distribution(combine_results, prob = 0.95)

}
